+++ packages/domain/src/index.ts
@@ -18,6 +18,9 @@ export {
   QuestionReviewStatusSchema,
   QuestionReviewAnswerSchema,
   AnswersRecordSchema,
+  AllowedPromptSchema,
+  PermissionModeSchema,
+  PlanReviewDecisionSchema,
 } from './entities/review'
 export type {
   ReviewStatus,
@@ -27,6 +30,9 @@ export type {
   QuestionReviewStatus,
   QuestionReviewAnswer,
   AnswersRecord,
+  AllowedPrompt,
+  PermissionMode,
+  PlanReviewDecision,
 } from './entities/review'
 
 export {
@@ -34,17 +40,20 @@ export {
   QuestionOptionSchema,
   QuestionItemSchema,
   QuestionRequestInputSchema,
+  PlanRequestInputSchema,
 } from './entities/hooks'
 export type {
   PermissionRequestInput,
   QuestionOption,
   QuestionItem,
   QuestionRequestInput,
+  PlanRequestInput,
 } from './entities/hooks'
 
 // Interfaces
 export type { IReviewRepository } from './repositories/IReviewRepository'
 export type { IQuestionReviewRepository } from './repositories/IQuestionReviewRepository'
+export type { IPlanReviewRepository } from './repositories/IPlanReviewRepository'
 export type { IHookTransport } from './adapters/IHookTransport'
 export type { IRoomResolver } from './services/IRoomResolver'
 
@@ -54,3 +63,6 @@ export type { ProcessPermissionReviewError } from './usecases/ProcessPermissionR
 
 export { default as ProcessQuestionReview } from './usecases/ProcessQuestionReview'
 export type { ProcessQuestionReviewError } from './usecases/ProcessQuestionReview'
+
+export { default as ProcessPlanReview } from './usecases/ProcessPlanReview'
+export type { ProcessPlanReviewError } from './usecases/ProcessPlanReview'